requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
av
beautifulsoup4
cachetools
certifi
charset-normalizer
click
colorama
coloredlogs
comtypes
contourpy
ctranslate2
cycler
distro
dnspython
docopt
fastapi
faster-whisper
filelock
filetype
flatbuffers
fonttools
frozenlist
fsspec
google-api-core
google-auth
google-cloud-texttospeech
googleapis-common-protos
greenlet
grpcio
grpcio-status
gTTS
h11
httpcore
httpx
huggingface-hub
humanfriendly
idna
Jinja2
jiter
Js2Py
kiwisolver
MarkupSafe
matplotlib
mpmath
multidict
networkx
numpy
onnxruntime
openai
opencv-python
opencv-python-headless
packaging
pandas
pillow
pillow_heif
propcache
proto-plus
protobuf
psutil
psycopg2-binary
py-cpuinfo
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pydub
pygame
pyjsparser
pymongo
pyparsing
PyPrind
pySmartDL
python-dateutil
python-dotenv
python-multipart
pyttsx3
pytz
PyYAML
regex
requests
requests-toolbelt
roboflow
rsa
safetensors
scipy
seaborn
setuptools
six
sniffio
soupsieve
SpeechRecognition
SQLAlchemy
starlette
sympy
tokenizers
torch
torchvision
tqdm
transformers
typing_extensions
tzdata
tzlocal
ultralytics
ultralytics-thop
urllib3
uvicorn
yarl
pyaudio
speechrecognition
